Postgis driver support (#1032)
* [pre-commit.ci] pre-commit autoupdate (#1022)

* [pre-commit.ci] pre-commit autoupdate

updates:
- [github.com/PyCQA/pylint: v3.1.0 → v3.2.3](https://github.com/PyCQA/pylint/compare/v3.1.0...v3.2.3)

* Lintage for Pylint upgrade.

* Non-pre-commit pylint too old.

---------


(cherry picked from commit bca8a8e205c3495f97122303f0bce38cc444aeec)

* Refactor run_sql to be shared code. Fix and document permissions behaviour for integration tests.

* Switch datacube-ows-update -e env option to -E env for consistency with datacube.

* Build postgis test database, including migrating geodata_coast to eo3.

* Add and register postgis index driver - no functionality implemented.

* Schema creation and management for postgis indexes.

* Postgis implementation, end-to-end testing and cleanup.

* Lintage/spelling and fix docker-compose for GH running.

* Attempt to get multi-db docker-compose test framework working in GHA.

* Fix pg_ready args.

* xargs was manging pg_isready args.

* Still trying to get pg_isready args passed correctly

* Ah need to replace DB_HOSTNAME.

* and typo.

* Update check-all-code.sh

* Separate probe db?

* Oh of course, datacube-ows-update can't be called until after we've schema-ed both dbs.

* Haha.  Oops.

* Core/OWS API fixes and workaround permissions bug in core.

* Lintage and mypy.

* Convert 4236 before clipping to CRS when generating CRS specific bboxes.

* MyPy fix.

* Fix time-query handling.

* Lintage.

* mypy fix.

* Fix postgis spatial extent method.

* Improve test coverage and remove unused code.

* Remove unused imports.

* Remove excess progress noise from datacube-ows-update --schema type commands.

* Update documentation.

* Silly spelling word.

* More spelling.

* Cleanup, add more comments.

* Cleanup, add more comments.

* Fix formatting of progress messages in run_sql().

* Fix sql.py copypasta.

---------

+++ datacube_ows/styles/component.py
@@ -165,7 +165,7 @@ class ComponentStyleDef(StyleDefBase):
                     else:
                         imgband_data = imgband_component
                 if imgband_data is None:
-                    null_np = np.zeros(tuple(data.sizes.values()), 'uint8')
+                    null_np = np.zeros(tuple(data.sizes.values()), 'float64')
                     imgband_data = DataArray(null_np, data.coords, tuple(data.sizes.keys()))
                 if imgband != "alpha":
                     imgband_data = self.compress_band(imgband, imgband_data)
